src/day-4/production/requirements.txt
agent-framework
azure-ai-agentserver-agentframework
python-dotenv
uvicorn
